backend/requirements.txt
passlib==1.7.4
pydantic==2.9.2
PyJWT==2.3.0
sqlmodel==0.0.22
mysql-connector-python==8.0.28
python-dotenv